Move SimpleNotification to core
