+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/util/SimpleNotification.kt
@@ -1,8 +1,8 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.core.util
 
 import io.gitlab.arturbosch.detekt.api.Notification
 
-data class SimpleNotification(
+internal data class SimpleNotification(
     override val message: String,
     override val level: Notification.Level = Notification.Level.Error
 ) : Notification {
